refactor(web): move shared boundary-drawing UI to components/hooks (tc-6he3x.14)
BoundaryMap, CustomShapeUpsell, and useBoundaryDrawing were reachable only
by importing across the WatchZones feature boundary, which the onboarding
flow (tc-6he3x.14) needs to reuse. react-coding-standards forbids
cross-feature imports, so relocate the three pieces to their proper shared
homes instead: BoundaryMap and CustomShapeUpsell to components/<Name>/
(matching ConfirmMap/RadiusPicker), useBoundaryDrawing to the flat hooks/
directory (matching useFetchData/usePagination). Pure move plus
import-path updates in WatchZoneCreatePage.tsx and WatchZoneEditPage.tsx;
no behavior change. useBoundaryDrawing's moved test now inlines its own
boundary literal instead of importing the WatchZones-feature fixture, to
avoid re-introducing the same isolation violation one level removed.

ShapeModeToggle stays in features/WatchZones/ - onboarding doesn't need
it (paid users go straight to the drawing step, no in-step toggle).
